CRC 64-bit fixes

+++ common/crc.cpp
@@ -58,7 +58,7 @@ void CRCEngine::operator()(char datum)
 {
     StagingBuffer.Buffer[Index++] = datum;
 
-    if (Index == sizeof(long)) {
+    if (Index == sizeof(int32_t)) {
         CRC = Value();
         StagingBuffer.Composite = 0;
         Index = 0;
@@ -83,7 +83,7 @@ void CRCEngine::operator()(char datum)
  * HISTORY:                                                                                    *
  *   03/02/1996 JLB : Created.                                                                 *
  *=============================================================================================*/
-long CRCEngine::operator()(void const* buffer, int length)
+int32_t CRCEngine::operator()(void const* buffer, long length)
 {
     if (buffer != NULL && length > 0) {
         char const* dataptr = (char const*)buffer;
@@ -105,11 +105,11 @@ long CRCEngine::operator()(void const* buffer, int length)
         **	Perform the fast 'bulk' processing by reading long word sized
         **	data blocks.
         */
-        long const* longptr = (long const*)dataptr;
-        int longcount = bytes_left / sizeof(long); // Whole 'long' elements remaining.
+        int32_t const* longptr = (int32_t const*)dataptr;
+        int longcount = bytes_left / sizeof(int32_t); // Whole 'long' elements remaining.
         while (longcount--) {
             CRC = lrotl(CRC, 1) + *longptr++;
-            bytes_left -= sizeof(long);
+            bytes_left -= sizeof(int32_t);
         }
 
         /*
@@ -147,7 +147,7 @@ long CRCEngine::operator()(void const* buffer, int length)
  * HISTORY:                                                                                    *
  *   03/02/1996 JLB : Created.                                                                 *
  *=============================================================================================*/
-long Calculate_CRC(void* buffer, long len)
+int32_t Calculate_CRC(void* buffer, long len)
 {
     return (CRCEngine()(buffer, len));
 }
